Port to sugar3.speech
sugar3.speech is used here as speech.py so the release of the activity wouldn't depend
on the release of the toolkit.

speech.py has been modified to include the changes in espeak.py that enabled mouth
movements, this file will be deleted once the file in the toolkit contains the changes
made here.

speechl.py is used here to import speech.py and has a function that gets the speech manager
from speech.py, it also contains the constants as used in espeak.py

mouth movement